% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRI.R
\name{calculateRI}
\alias{calculateRI}
\title{Calculates the non-isothermal Kovats retention indices (from temperature-programming, using definition of Van den Dool and Kratz: Ix = 100n + 100(tx-tn) / (tn+1 − tn)}
\usage{
calculateRI(data, alkaneSeries)
}
\arguments{
\item{data}{data.frame with name and rt (in minutes)}

\item{alkaneSeries}{data.frame with Num (number of carbons in of the alkane) and rt (in minutes)}
}
\value{

}
\description{
Calculates the non-isothermal Kovats retention indices (from temperature-programming, using definition of Van den Dool and Kratz: Ix = 100n + 100(tx-tn) / (tn+1 − tn)
}
