% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMSP.R
\name{exportMSP}
\alias{exportMSP}
\title{Export to NIST msp format
Code modified from Stettin et al. doi: 10.3390/metabo10040143}
\usage{
exportMSP(data, out.file, min.clustersize = 15, num.files, alkaneSeries = NULL)
}
\arguments{
\item{data}{an XSAnnotate object generated from CAMERA}

\item{out.file}{character name of the output msp file (.msp will be automatically added)}

\item{min.clustersize}{numeric how many fragments need to be in a group for it to be included in the results. 5 - trace compounds; 20 - mostly high quality spectra}

\item{num.files}{numeric (length = 1) specify how many files are included in the analysis}

\item{alkaneSeries}{}
}
\value{

}
\description{
Export to NIST msp format
Code modified from Stettin et al. doi: 10.3390/metabo10040143
}
\examples{
alkaneSeries <- data.frame(Cnum = c(11, 13, 15, 17, 19, 21, 23, 25), rt = c(4.90, 8.00, 11.13, 14.05, 16.70, 19.37, 22.45, 25.77))

}
