% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_homolog_MDPlot.R
\name{find_homologs_MDPlot}
\alias{find_homologs_MDPlot}
\title{find_homologs_MDPlot
A wrapper for homol.search and plothomol from the nontarget package with explorative output using plotly. Default values are for PFAS search.}
\usage{
find_homologs_MDPlot(
  file,
  plotdefect = FALSE,
  intensity = intensity,
  p_elements = c("C", "H", "O", "S", "F", "N"),
  p_use_C = FALSE,
  p_minmz = 49.9,
  p_maxmz = 50,
  p_minrt = 0.5,
  p_maxrt = 2,
  p_ppm = TRUE,
  p_mztol = 12,
  p_rttol = 0.5,
  p_minlength = 3,
  p_mzfilter = FALSE,
  p_vec_size = 3e+06,
  p_mat_size = 3,
  p_R2 = 0.98,
  p_spar = 0.45,
  p_plotit = FALSE,
  p_deb = 0
)
}
\arguments{
\item{file}{a csv file containing 'mz', 'intensity' and 'rt' as variable names. 'rt' should be in minutes.}

\item{plotdefect}{logical. Whether or not to plot mass defect (mz-round(mz)) instead of rt. Defaults to FALSE.}

\item{intensity}{variable name. If multiple sample intensities are in the dataframe, then you can choose the specific sample intensity. Defaults to 'intensity'.}

\item{p_elements}{character vector. FALSE or chemical elements in the changing units of the homologue series, e.g. c("C","H") for alkane chains or c("C", "F") for perfluorinated compounds. Used to restrict search.Elements to include in the homolog search. Defaults to: c("C", "H", "O")}

\item{p_use_C}{logical. For elements: take element ratio to C-atoms into account? Used to restrict search}

\item{p_minmz}{Defines the lower limit of the m/z window to search homologue series peaks, relative to the m/z of the one peak to search from. Absolute m/z value [u].}

\item{p_maxmz}{Defines the upper limit of the m/z window to search homologue series peaks, relative to the m/z of the one peak to search from. Absolute m/z value [u].}

\item{p_minrt}{Defines the lower limit of the retention time (RT) window to look for other homologue peaks, relative to the RT of the one peak to search from, i.e., RT+minrt. For decreasing RT with increasing HS mass, use negative values of minrt.}

\item{p_maxrt}{Defines the upper limit of the RT window to look for other homologue peaks, relative to the RT of the one peak to search from, i.e., RT+maxrt. See minrt.}

\item{p_ppm}{Should mztol be set in ppm (TRUE) or in absolute m/z [u] (FALSE)?}

\item{p_mztol}{m/z tolerance setting: +/- value by which the m/z of a peak may vary from its expected value. If parameter ppm=TRUE (see below) given in ppm, otherwise, if ppm=FALSE, in absolute m/z [u].}

\item{p_rttol}{Retention time (RT) tolerance by which the RT between two adjacent pairs of a homologue series is allowed to differ. Units as given in column 3 of peaklist argument, e.g. [min].}

\item{p_minlength}{Minimum number of peaks in a homologue series.}

\item{p_mzfilter}{Vector of numerics to filter for homologue series with specific m/z differences of their repeating units, given the tolerances in mztol. Mind charge z!}

\item{p_vec_size}{Vector size. Ignore unless a relevant error message is printed (then try to increase size).}

\item{p_mat_size}{Matrix size for recombining, multiple of input tuples. Ignore unless a relevant error message is printed (then try to increase size).}

\item{p_R2}{FALSE or 0<numeric<=1. Coefficient of determination for cubic smoothing spline fits of m/z versus retention time; homologue series with lower R2 are rejected. See smooth.spline.}

\item{p_spar}{Smoothing parameter, typically (but not necessarily) in (0,1]. See smooth.spline.}

\item{p_plotit}{Logical FALSE or 0<integer<5. Intermediate plots of nearest neighbor paths, spline fits of individual homologues series >=minlength, clustered HS pairs, etc .}

\item{p_deb}{Debug returns, ignore.}
}
\value{

}
\description{
find_homologs_MDPlot
A wrapper for homol.search and plothomol from the nontarget package with explorative output using plotly. Default values are for PFAS search.
}
\examples{
find_homologs("./data/LCneg_ComponentsRC.csv", 
plotdefect = FALSE,
intensity = N_rep1,
p_elements=c("C","H", "O", "S", "F"),
p_use_C=FALSE,
p_minmz=49.9,
p_maxmz=50,
p_minrt=0.5,
p_maxrt=2,
p_ppm=TRUE,
p_mztol=10,
p_rttol=0.5,
p_minlength = 3,
p_mzfilter=FALSE,
p_vec_size=3E6,
p_mat_size=3,
p_R2=.98,
p_spar=.45,
p_plotit=FALSE,
p_deb=0)
}
