% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormulaFromMSP.R
\name{getFormulaFromMSP}
\alias{getFormulaFromMSP}
\title{getFormulaFromMSO}
\usage{
getFormulaFromMSP(file)
}
\arguments{
\item{file}{}
}
\value{
character vector of chemical formula
}
\description{
getFormulaFromMSO
}
