% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmass.R
\name{getmass}
\alias{getmass}
\title{getmass - helper function to MDPlotR
Get the exact mass from chemical formula using rcdk}
\usage{
getmass(data)
}
\arguments{
\item{data}{}
}
\value{
vector
}
\description{
getmass - helper function to MDPlotR
Get the exact mass from chemical formula using rcdk
}
