% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmdh.R
\name{getmdh}
\alias{getmdh}
\title{Mass defect calculations - helper function to MDPlotR}
\usage{
getmdh(mz, cus = c("CH2,H2"), method = "round")
}
\arguments{
\item{mz}{}

\item{cus}{}

\item{method}{}
}
\value{
data.frame
}
\description{
Mass defect calculations - helper function to MDPlotR
}
