% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRMF_all.R
\name{HRMF_all}
\alias{HRMF_all}
\title{HRMF: High resolution mass filtering for GC HRMS data for multiple entries in a msp file}
\usage{
HRMF_all(
  file,
  charge = 1,
  mass_accuracy = 5,
  intensity_cutoff = 1,
  IR_RelAb_cutoff = 1
)
}
\arguments{
\item{file}{}

\item{mass_accuracy}{integer The mass accuracy in ppm}

\item{intensity_cutoff}{integer  The absolute intensity cutoff stated in the msp}

\item{IR_RelAb_cutoff}{integer The relative abundance cutoff for the isotopologues relative to the monoisotopic mass}
}
\value{
list   list of outputs consisting of dataframes
}
\description{
HRMF: High resolution mass filtering for GC HRMS data for multiple entries in a msp file
}
\examples{
hit <- HRMF(file = "D://R_projects//MSxplorer//data//2-Methoxy-5-methylaniline_C8H11NO.msp", formula = "C8H11NO", IR_RelAb_cutoff = 1, intensity_cutoff = 100000)
# TO CHECK
High resolution mass filtering for GC HRMS data
}
