% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEIC.R
\name{plotEIC}
\alias{plotEIC}
\title{Plot extracted ion chromatograms}
\usage{
plotEIC(filepath, featlist, diff = 0.005, mserr = 0.005)
}
\arguments{
\item{filepath}{Path to the raw data file(s)}

\item{featlist}{A data frame containing the columns: name, mz, ms_level}

\item{diff}{A single value specifying the mass error}

\item{mserr}{The mass error in Dalton}
}
\value{
Overlayed extracted chromatograms
}
\description{
Plot the EIC from MS1 or MS2 from raw data using MSnbase
}
