% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMS1Peaks.R
\name{plotTopMS1Peaks}
\alias{plotTopMS1Peaks}
\title{Extract most intense MS1 ions from MS2 EIC using Shiny}
\usage{
plotTopMS1Peaks(
  filepath,
  featlist,
  numTopIons = 10,
  diff = 0.01,
  mserr = 0.01,
  rtWindow = 0.3
)
}
\arguments{
\item{filepath}{string. Path to mzML file}

\item{featlist}{data.frame. Variable names should be: name, mz, ms_level (has to be either "ms1, "ms2)}

\item{numTopIons}{num number of most intensive MS1 ions to extract}

\item{diff}{num range of mass (Da) to extract from the mz specified in featlist (mz-diff, mz+diff)}

\item{mserr}{num mass error (Da) of the range boundaries specified in diff (mz-diff+-mserr, mz+dif+-mserr)}

\item{rtWindow}{num the retention time range to extract MS1 EIC (in sec). Error occurs if set too low.}
}
\value{
overlaid EIC from MS2 and MS1 in an interactive Shiny environment
}
\description{
Extract most intense MS1 ions from MS2 EIC using Shiny
}
\examples{
fl <- "D:/OneDrive - Linköpings universitet/Raw_data/Dust_Florian/LC/LC_neg_mzML"

PFSA_frags <- data.frame(name = c("FSO3", "SO3"), mz = c(98.9552, 79.9558), ms_level = c("ms2", "ms2"))

plotTopMS1Peaks(filepath = fl, featlist = PFSA_frags, numTopIons = 3)

fl <- "D:/OneDrive - Linköpings universitet/Raw_data/Dust_Florian/LC/LC_neg_mzMLQC_rep1.mzML"
Br_frags <- data.frame(name = c("Br79", "Br81"), mz = c(78.9183, 80.9163), ms_level = c("ms2", "ms2"))
plotTopMS1Peaks(filepath = fl, featlist = Br_frags, numTopIons = 3)
}
