% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMS2Peaks.R
\name{plotTopMS2Peaks}
\alias{plotTopMS2Peaks}
\title{Extract most intense MS2 ions from MS1 EIC using Shiny}
\usage{
plotTopMS2Peaks(
  filepath,
  featlist,
  numTopIons = 10,
  diff = 0.01,
  mserr = 0.01,
  rtWindow = 0.3
)
}
\arguments{
\item{filepath}{string. Path to mzML/mzXML file}

\item{featlist}{data.frame. Variable names should be: name, mz, ms_level (has to be either "ms1, "ms2)}

\item{numTopIons}{num number of most intensive MS2 ions to extract}

\item{diff}{num range of mass (Da) to extract from the mz specified in featlist (mz-diff, mz+diff)}

\item{mserr}{num mass error (Da) of the range boundaries specified in diff (mz-diff+-mserr, mz+dif+-mserr)}

\item{rtWindow}{num the retention time range to extract MS2 EIC (in sec). Error occurs if set too low.}
}
\value{
overlaid EIC from MS2 and MS1 in an interactive Shiny environment
}
\description{
Extract most intense MS2 ions from MS1 EIC using Shiny
}
\examples{

}
