% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msp.R
\name{read_msp}
\alias{read_msp}
\title{read_msp}
\usage{
read_msp(file)
}
\arguments{
\item{file, }{a string with the path of the msp file in either NIST or RIKEN format}
}
\value{
a list of dataframes from individual compounds in the msp file containing mz, intensity and annotation (if available)
}
\description{
read_msp
}
\examples{
file <- "./data/GC_orbitrapSTD_RIKEN.msp"
peak_table_list <- read_msp(file)
}
