% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMSP.R
\name{exportMSP}
\alias{exportMSP}
\title{Export to NIST msp format}
\usage{
exportMSP(data, RTI = NULL)
}
\arguments{
\item{data}{dataframe}

\item{RTI}{dataframe containing: CNum (carbon number of the alkane) and RT (retention time in minutes).}
}
\value{

}
\description{
Export to NIST msp format
}
\examples{

}
