% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmass.R
\name{getmass}
\alias{getmass}
\title{getmass
Get the exact mass from chemical formula using rcdk}
\usage{
getmass(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
getmass
Get the exact mass from chemical formula using rcdk
}
