% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMS1Peaks.R
\name{plotTopMS1Peaks}
\alias{plotTopMS1Peaks}
\title{Extract most intense MS1 ions from MS2 EIC}
\usage{
plotTopMS1Peaks(
  filepath,
  flagfragments,
  numTopIons = 10,
  diff = 0.01,
  mserr = 0.01,
  rtrange = 0.5
)
}
\arguments{
\item{filepath}{string path to mzML file}

\item{flagfragments}{data.frame with variables: name, mz, ms_level ("ms1, "ms2)}

\item{numTopIons}{num number of most intensive MS1 ions to extract}

\item{diff}{num range of mass (Da) to extract from the mz specified in flagfragments (mz-diff, mz+diff)}

\item{mserr}{num mass error (Da) of the range boundaries specified in diff (mz-diff+-mserr, mz+dif+-mserr)}

\item{rtrange}{num the retention time range to extract MS1 EIC (in sec). Error occurs if set too low.}
}
\value{
overlaid EIC from MS2 and MS1
}
\description{
Extract most intense MS1 ions from MS2 EIC
}
\examples{
fl <- "D:/Raw_data/Kallinge/New_analysis_20200414/centroid/B6 batch std_1_F,2_1.mzML"
PFSA_frags <- data.frame(name = c("FSO3", "SO3"), mz = c(98.9552, 79.9558), ms_level = c("ms2", "ms2"))
plotTopMS1Peaks(filepath = fl, flagfragments = PFSA_frags, numTopIons = 3)
}
